+++ src/Blog/elements/PostAbstract/PostAbstract.component.tsx
@@ -2,6 +2,6 @@ import * as markdown from '../../../Shared/lib/markdown';
 
 import { MarkdownComponentProps, markdownFactory } from '../../../Shared/elements/Markdown/Markdown.factory';
 
-const createComponent = () => markdownFactory(markdown.strippedLinks(), {});
+const createComponent = () => markdownFactory(markdown.strippedLinks(), {}, 'post-abstract');
 
 export const PostAbstract: React.StatelessComponent<MarkdownComponentProps> = createComponent();