+++ src/Blog/templates/Post/PostTemplate.component.tsx
@@ -4,6 +4,9 @@ import { Head } from 'react-static';
 import { Link } from '../../../Shared/elements/Link/Link.component';
 import { DefaultLayout as Layout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
 
+import { SiteFooter } from '../../../Site/blocks/SiteFooter/SiteFooter.component';
+import { SiteHeader } from '../../../Site/blocks/SiteHeader/SiteHeader.component';
+
 import { PostAbstract } from '../../elements/PostAbstract/PostAbstract.component';
 import { PostContents } from '../../elements/PostContents/PostContents.component';
 import { Post } from '../../types/Post.model';
@@ -14,14 +17,30 @@ interface Props {
   post: Post;
 }
 
-export const PostTemplate: React.StatelessComponent<Props> = props => {
+export const PostTemplate: React.StatelessComponent<Props> = ({ post }) => {
+  const subHeader = (
+    <header>
+      <Link id="aria-page-title" to="/posts/">
+        Blog
+      </Link>
+      <Link id="aria-page-title" to="/posts/">
+        Next Post
+      </Link>
+      <Link id="aria-page-title" to="/posts/">
+        Previous Post
+      </Link>
+    </header>
+  );
+  const header = <SiteHeader page={post}>{subHeader}</SiteHeader>;
+  const footer = <SiteFooter />;
   return (
-    <Layout>
-      <Head title={props.post.title} meta={props.post.meta} />
-      <Link to="/posts/">{'<'} Back</Link>
-      <h1 className={styles.Post}>{props.post.title}</h1>
-      <PostAbstract content={props.post.abstract} />
-      <PostContents content={props.post.content} />
+    <Layout className="blog-post" header={header} footer={footer}>
+      <Head title={post.title} meta={post.meta} />
+      <article className={`post ${styles.Module}`}>
+        <h1 className={styles.Title}>{post.title}</h1>
+        <PostAbstract>{post.abstract}</PostAbstract>
+        <PostContents>{post.content}</PostContents>
+      </article>
     </Layout>
   );
 };