+++ src/Docs/templates/Index/IndexTemplate.component.tsx
@@ -2,7 +2,10 @@ import * as React from 'react';
 import { Head } from 'react-static';
 
 import { DefaultLayout as Layout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
-import { ContentPage, PageRouteData } from '../../../Shared/types/Page.model';
+import { ContentPage } from '../../../Shared/types/Page.model';
+
+import { SiteFooter } from '../../../Site/blocks/SiteFooter/SiteFooter.component';
+import { SiteHeader } from '../../../Site/blocks/SiteHeader/SiteHeader.component';
 
 import { DocContents } from '../../elements/DocContents/DocContents.component';
 
@@ -10,12 +13,14 @@ interface Props {
   page: ContentPage;
 }
 
-export const IndexTemplate: React.StatelessComponent<Props> = props => {
+export const IndexTemplate: React.StatelessComponent<Props> = ({ page }) => {
+  const header = <SiteHeader page={page} />;
+  const footer = <SiteFooter />;
   return (
-    <Layout>
-      <Head title={props.page.title} meta={props.page.meta} />
-      <h1>{props.page.title}</h1>
-      <DocContents content={props.page.content} />
+    <Layout className="docs-index" header={header} footer={footer}>
+      <Head title={page.title} meta={page.meta} />
+      <h1>{page.title}</h1>
+      <DocContents>{page.content}</DocContents>
     </Layout>
   );
 };