+++ src/Shared/behaviours/AutoScrollToTop/AutoScrollToTop.component.tsx
@@ -1,5 +1,6 @@
-import { Location } from '@reach/router';
-import React, { FunctionComponent } from 'react';
+import * as React from 'react';
+
+// props: https://gist.github.com/swernerx/2c2ba4e611b4ec7921813a71517ddf5a
 
 let lastNavigationFromBrowserUI = true;
 
@@ -9,29 +10,20 @@ if (typeof (window as any) !== 'undefined') {
   });
 }
 
-// props: https://gist.github.com/swernerx/2c2ba4e611b4ec7921813a71517ddf5a
+interface Props {
+  children?: React.ReactNode;
+}
 
-export const AutoScrollToTop: FunctionComponent = ({ children }) => (
-  <Location>
-    {() => {
-      if (typeof (history as any) !== 'undefined') {
-        // Ininitial rendering and back/forward navigation uses browsers
-        // native scroll history mechanism which tracks scroll position
-        // for each history entry automatically
-        if (lastNavigationFromBrowserUI) {
-          lastNavigationFromBrowserUI = false;
-        } else {
-          // When adding new entries by navigating through clicking on actual
-          // links in the page, we always scroll up to work around
-          // the scrolling applied by automatic focussing as done
-          // by reach routers accessibility tweaks.
-          requestAnimationFrame(() => {
-            window.scrollTo(0, 0);
-          });
-        }
-      }
+export class AutoScrollToTop extends React.Component<Props> {
+  public componentDidMount() {
+    if (lastNavigationFromBrowserUI) {
+      lastNavigationFromBrowserUI = false;
+    } else {
+      window.scrollTo(0, 0);
+    }
+  }
 
-      return children;
-    }}
-  </Location>
-);
+  public render() {
+    return this.props.children;
+  }
+}