+++ src/Shared/elements/Link/Link.component.tsx
@@ -1,7 +1,8 @@
-import { Link as RouterLink } from '@reach/router';
 import * as React from 'react';
+import { Link as RouterLink } from 'react-router-dom';
 
 const isExternal = (url: string): boolean => /^[a-z]+:/.test(url);
+const isFragment = (url: string): boolean => /^#.+/.test(url);
 
 interface Props {
   href?: string;
@@ -15,6 +16,7 @@ const external = (url: string, id: string, className: string, children?: React.R
   const props = { href: url, id, className };
   return <a {...props}>{children}</a>;
 };
+
 const internal = (url: string, id: string, className: string, children?: React.ReactNode) => {
   const props = { to: url, id, className };
   return <RouterLink {...props}>{children}</RouterLink>;
@@ -22,6 +24,6 @@ const internal = (url: string, id: string, className: string, children?: React.R
 
 export const Link: React.StatelessComponent<Props> = ({ href, to, id, className, children }) => {
   const url = href || to || '/';
-  const render = isExternal(href) ? external : internal;
+  const render = isExternal(href) || isFragment(href) ? external : internal;
   return render(url, id, className, children);
 };