+++ src/Site/templates/ErrorNotFound/NotFoundTemplate.component.tsx
@@ -4,7 +4,10 @@ import { Head } from 'react-static';
 import { DefaultLayout as Layout } from '../../../Shared/layout/DefaultLayout/DefaultLayout.component';
 // import { ContentPage } from '../../../Shared/types/Page.model';
 
-import * as styles from './NotFoundTemplate.module.scss';
+import { SiteFooter } from '../../../Site/blocks/SiteFooter/SiteFooter.component';
+import { SiteHeader } from '../../../Site/blocks/SiteHeader/SiteHeader.component';
+
+// import * as styles from './NotFoundTemplate.module.scss';
 
 // interface Props {
 //   page: ContentPage;
@@ -16,8 +19,10 @@ export const NotFoundTemplate: React.StatelessComponent<{}> = () => {
       title: '404 Not Found'
     }
   };
+  const header = <SiteHeader page={props.page} />;
+  const footer = <SiteFooter />;
   return (
-    <Layout className={styles.layout}>
+    <Layout className="error-not-found" header={header} footer={footer}>
       <Head title={props.page.title} />
       <h1>{props.page.title}</h1>
       {/* <h1 className="title-page">Latest Posts {posts.length}</h1> */}